+++ libpod/container_inspect.go
@@ -151,7 +151,7 @@ func (c *Container) getContainerInspectData(size bool, driverData *define.Driver
 
 	if c.config.HealthCheckConfig != nil {
 		// This container has a healthcheck defined in it; we need to add it's state
-		healthCheckState, err := c.GetHealthCheckLog()
+		healthCheckState, err := c.getHealthCheckLog()
 		if err != nil {
 			// An error here is not considered fatal; no health state will be displayed
 			logrus.Error(err)