+++ pkg/domain/infra/abi/containers_runlabel.go
@@ -133,6 +133,9 @@ func generateRunlabelCommand(runlabel string, img *libimage.Image, inputName str
 		}
 		splitImageName := strings.Split(normalize, "/")
 		name = splitImageName[len(splitImageName)-1]
+		// make sure to remove the tag from the image name, otherwise the name cannot
+		// be used as container name because a colon is an illegal character
+		name = strings.SplitN(name, ":", 2)[0]
 	}
 
 	// Append the user-specified arguments to the runlabel (command).